update module names per pep8 style guide
