+++ example/example_script.py
@@ -11,7 +11,7 @@ CONDITIONS OF ANY KIND, either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 """
 
-from floris.Floris import Floris
+from floris.floris import Floris
 
 floris = Floris("floris.json")
