+++ floris/base_object.py
