+++ floris/visualization_manager.py
@@ -11,7 +11,7 @@ CONDITIONS OF ANY KIND, either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 """
 
-from .Coordinate import Coordinate
+from .coordinate import Coordinate
 import matplotlib.pyplot as plt
 import numpy as np
