+++ floris/wake_combination.py
@@ -11,7 +11,7 @@ CONDITIONS OF ANY KIND, either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 """
 
-from .BaseObject import BaseObject
+from .base_object import BaseObject
 import numpy as np
