+++ tests/floris_unit.py
@@ -11,14 +11,17 @@ CONDITIONS OF ANY KIND, either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 """
 
-from floris.BaseObject import BaseObject
+from .base_object_test import BaseObjectTest
+from .coordinate_test import CoordinateTest
+from .flow_field_test import FlowFieldTest
 
-class BaseObjectTest():
-    def test_all(self):
-        test_object_instantiation()
+class FlorisUnitTest():
+    def __init__(self):
+        self.baseobject = BaseObjectTest()
+        self.coordinate = CoordinateTest()
+        self.flowfield = FlowFieldTest()
 
-def test_object_instantiation():
-    """
-    The class should initialize with the standard inputs
-    """
-    assert BaseObject() is not None
+    def run_tests(self):
+        self.baseobject.test_all()
+        self.coordinate.test_all()
+        self.flowfield.test_all()
